pyproject.toml
[project]
name = "kratos"
version = "0.1.0"
description = "Cybersecurity AI Terminal Agent for CTFs and Pentesting"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{ name = "Vidar" }]
dependencies = [
    "ollama>=0.4.0",
    "pydantic>=2.10,<3",
    "rich>=13.9.0",
    "prompt-toolkit>=3.0.39",
    "docker>=7.0.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
tui = ["textual>=0.50.0"]
dev = [
    "ruff>=0.9.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "mypy>=1.10",
]

[project.scripts]
kratos = "kratos.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kratos"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
